# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Article:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      category_id:
        type: integer
      category_name:
        type: string
      content:
        description: omitted in list, included in detail
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      published_at:
        type: string
      read_time_minutes:
        type: integer
      slug:
        type: string
      status:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  domain.AuthResponse:
    properties:
      access_expiry:
        type: string
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  domain.CreateArticleRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      description:
        type: string
      image_url:
        type: string
      status:
        description: '"draft" or "published"'
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.NewsFeedResponse:
    properties:
      feed_articles:
        items:
          $ref: '#/definitions/domain.Article'
        type: array
      hero_article:
        $ref: '#/definitions/domain.Article'
      meta:
        $ref: '#/definitions/domain.PaginationMeta'
    type: object
  domain.PaginationMeta:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  domain.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  domain.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  domain.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        type: string
      preferences:
        type: string
    required:
    - name
    type: object
  domain.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      preferences:
        description: JSON encoded string for simplicity, or json.RawMessage
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  util.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: 103.181.143.73:8081
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend REST API for a Flutter News App with Auth + News Feature
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Free API News
  version: "1.0"
paths:
  /admin/articles:
    post:
      consumes:
      - application/json
      description: Creates a news article. Slug is auto-generated from the title.
        read_time_minutes is auto-calculated (200 words/min). published_at is set
        automatically when status=published.
      parameters:
      - description: Article payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: article created
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Create a new article (Admin)
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get tokens
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: login successful
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token
      parameters:
      - description: Logout Request (refresh token to revoke)
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: logged out successfully
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Returns user info based on validated JWT
      produces:
      - application/json
      responses:
        "200":
          description: authenticated user
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
    put:
      consumes:
      - application/json
      description: Updates authenticated user profile data
      parameters:
      - description: Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: profile updated
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access and refresh token pair using an unexpired refresh
        token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token refreshed
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
      summary: Refresh token pair
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register with name, email and password (min 8 chars)
      parameters:
      - description: Register Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: registration successful
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: Register a new user
      tags:
      - auth
  /news:
    get:
      description: Returns a paginated news feed. When include_hero=true (default),
        the first article on page 1 is separated as hero_article for the Flutter Hero+Grid
        UI. Supports infinite scroll via the page query param.
      parameters:
      - description: Filter by category slug (e.g. 'technology', 'sports')
        in: query
        name: category
        type: string
      - default: 1
        description: Page number, starts at 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page, max 50
        in: query
        name: limit
        type: integer
      - default: true
        description: Separate first article as hero
        in: query
        name: include_hero
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: paginated feed
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.NewsFeedResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Get news feed
      tags:
      - news
  /news/{slug}:
    get:
      description: Returns the full content of a single article by its URL slug
      parameters:
      - description: Article slug (e.g. 'apples-ai-leap-iphone-16-pro')
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: article detail
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Get article detail
      tags:
      - news
  /news/categories:
    get:
      description: Returns all active news categories available as filter chips in
        the feed UI
      produces:
      - application/json
      responses:
        "200":
          description: list of active categories
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Category'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: List active categories
      tags:
      - news
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file (multipart/form-data) and returns its public
        URL
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  properties:
                    url:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: Upload an image
      tags:
      - upload
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"