# internal/domain/user.go
package domain

import "time"

// User is the core domain entity
type User struct {
	ID          int64     `json:"id"`
	Name        string    `json:"name"`
	Email       string    `json:"email"`
	Password    string    `json:"-"` // never expose
	Role        string    `json:"role"`
	AvatarURL   string    `json:"avatar_url"`
	Bio         string    `json:"bio"`
	Phone       string    `json:"phone"`
	Preferences string    `json:"preferences"` // JSON encoded string for simplicity, or json.RawMessage
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}

// Token represents an auth token pair
type Token struct {
	ID           int64     `json:"id"`
	UserID       int64     `json:"user_id"`
	AccessToken  string    `json:"access_token"`
	RefreshToken string    `json:"refresh_token"`
	AccessExpiry time.Time `json:"access_expiry"`
	RefreshExpiry time.Time `json:"refresh_expiry"`
	IsRevoked    bool      `json:"is_revoked"`
	CreatedAt    time.Time `json:"created_at"`
}

// RegisterRequest is the input for sign-up
type RegisterRequest struct {
	Name     string `json:"name"     validate:"required,min=2,max=100"`
	Email    string `json:"email"    validate:"required,email"`
	Password string `json:"password" validate:"required,min=8"`
}

// LoginRequest is the input for sign-in
type LoginRequest struct {
	Email    string `json:"email"    validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

// AuthResponse is returned after a successful register/login
type AuthResponse struct {
	User         *User     `json:"user"`
	AccessToken  string    `json:"access_token"`
	RefreshToken string    `json:"refresh_token"`
	AccessExpiry time.Time `json:"access_expiry"`
}

// RefreshRequest is the input for refreshing tokens
type RefreshRequest struct {
	RefreshToken string `json:"refresh_token" validate:"required"`
}

// UpdateProfileRequest is the input for updating user profile
type UpdateProfileRequest struct {
	Name        string `json:"name" validate:"required,min=2,max=100"`
	AvatarURL   string `json:"avatar_url"`
	Bio         string `json:"bio"`
	Phone       string `json:"phone"`
	Preferences string `json:"preferences"`
}